rename backIcon to BackIcon
